chore: rename template
