+++ packages/@sanity/cli/src/actions/init-project/templates/appSanityUi.ts
