+++ packages/@sanity/cli/src/actions/init-project/templates/index.ts
@@ -1,6 +1,6 @@
 import {type ProjectTemplate} from '../initProject'
 import appTemplate from './appQuickstart'
-import appSanityUiTemplate from './appQuickstartSanityUi'
+import appSanityUiTemplate from './appSanityUi'
 import blog from './blog'
 import clean from './clean'
 import getStartedTemplate from './getStarted'
@@ -13,7 +13,7 @@ const templates: Record<string, ProjectTemplate | undefined> = {
   blog,
   clean,
   'app-quickstart': appTemplate,
-  'app-quickstart-sanity-ui': appSanityUiTemplate,
+  'app-sanity-ui': appSanityUiTemplate,
   'get-started': getStartedTemplate,
   moviedb,
   shopify,
