+++ packages/@sanity/cli/templates/app-sanity-ui/src/App.tsx
