+++ packages/@sanity/cli/templates/app-sanity-ui/src/ExampleComponent.tsx
